pyproject.toml
[project]
name = "kanban_board"
version = "0.1.0"
description = ""
authors = [
    {name = "Max",email = "maks_artyukhin@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2,<6.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "drf-yasg (>=1.21.10,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.1.0"
black = "^25.1.0"
flake8 = "^7.2.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["users", "phone", "core", "kanban_board"]
